requirements.txt
Package         Version
--------------- -------
greenlet        1.1.2
pip             21.3.1
python-decouple 3.5
setuptools      57.4.0
requests~=2.27.1
google-api-python-client~=2.41.0
google-auth-oauthlib==0.5.1
termcolor~=1.1.0